update becnhmarks and rename few things
