+++ benchmark-results.json
@@ -1,1 +1,1 @@
-[{"name":"http-bare","version":"10.13.0","hasRouter":false,"requests":"18045.3","latency":"54.87","throughput":"4.63","validation":"✗","description":"Super basic and completely useless bare http server, should be the theoretical upper limit in performance."},{"name":"fastify","version":"4.9.2","hasRouter":true,"requests":"15201.8","latency":"65.24","throughput":"3.91","validation":"-","description":"Validation is done using schemas and ajv. Schemas must be generated manually or using third party tools."},{"name":"mion","version":"0.1.0","hasRouter":true,"requests":"12872.8","latency":"77.16","throughput":"3.57","validation":"✓","description":"Automatic validation out of the box using @deepkit/types."},{"name":"restify","version":"8.6.1","hasRouter":true,"requests":"11215.9","latency":"88.62","throughput":"2.90","validation":"✗","description":"Requires third party tools."},{"name":"hapi","version":"20.2.2","hasRouter":true,"requests":"7886.4","latency":"126.13","throughput":"2.02","validation":"✗","description":"Manual validation using joi, or third party tools."},{"name":"express","version":"4.18.2","hasRouter":true,"requests":"4600.7","latency":"216.34","throughput":"1.18","validation":"✗","description":"needs third party tools, or third party tools"},{"name":"deepkit","version":"1.0.1-alpha.75","hasRouter":true,"requests":"1703.9","latency":"583.36","throughput":"0.44","validation":"✓","description":"Automatic validation out of the box (The ones that made @deepkit/types👍). They have a RPC over webSockets that's way more performant than the http tested here."}]
+[{"name":"node-http","version":"16.18.0","hasRouter":false,"requests":"19594.5","latency":"50.50","throughput":"5.03","validation":"✗","description":"Super basic and completely useless bare http server, should be the theoretical upper limit in performance."},{"name":"fastify","version":"4.9.2","hasRouter":true,"requests":"16831.4","latency":"58.88","throughput":"4.33","validation":"-","description":"Validation is done using schemas and ajv. Schemas must be generated manually or using third party tools."},{"name":"mion","version":"0.1.0","hasRouter":true,"requests":"13150.6","latency":"75.49","throughput":"3.65","validation":"✓","description":"Automatic validation out of the box using @deepkit/types."},{"name":"restify","version":"8.6.1","hasRouter":true,"requests":"11766.4","latency":"84.41","throughput":"3.04","validation":"✗","description":"Requires third party tools."},{"name":"hapi","version":"20.2.2","hasRouter":true,"requests":"7173.4","latency":"138.40","throughput":"1.84","validation":"✗","description":"Manual validation using joi, or third party tools."},{"name":"express","version":"4.18.2","hasRouter":true,"requests":"4497.8","latency":"221.34","throughput":"1.15","validation":"✗","description":"needs third party tools, or third party tools"},{"name":"deepkit","version":"1.0.1-alpha.75","hasRouter":true,"requests":"1647.8","latency":"604.28","throughput":"0.42","validation":"✓","description":"Automatic validation out of the box (The ones that made @deepkit/types), Their rpc is way more performant."}]
