+++ benchmarks/deepkit.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { initDeepkitApp, setRoutes } = require("../compiled-apps/apps");
+const { initDeepkitApp, setRoutes } = require("../_compiled-apps/apps");
 
 // ###### check the apps/ directory for the original non compiled code
 // deepkit needs to be compiled from typescript to be able to generate runtime types metadata
