+++ benchmarks/mion.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { initHttp, addRoutes, routes } = require("../compiled-apps/apps");
+const { initHttp, addRoutes, routes } = require("../_compiled-apps/apps");
 const { startHttpServer } = require("@mionkit/http");
 
 // ###### check the apps/ directory for the original non compiled code
