+++ lib/packages.js
@@ -3,14 +3,14 @@
 const { dependencies } = require("../package");
 
 const packages = {
-  "http-bare": {
+  "node-http": {
     hasRouter: false,
-    version: "10.13.0",
+    version: "16.18.0",
     validation: "✗",
     description:
       "Super basic and completely useless bare http server, should be the theoretical upper limit in performance.",
   },
-  // "mion-http-bare": {
+  // "mion-node-http": {
   //   hasRouter: false,
   //   package: "@MionKit/compiled-app",
   //   validation: "✗",
@@ -38,7 +38,7 @@ const packages = {
     validation: "✓",
     version: "1.0.1-alpha.75",
     description:
-      "Automatic validation out of the box (The ones that made @deepkit/types👍). They have a RPC over webSockets that's way more performant than the http tested here.",
+      "Automatic validation out of the box (The ones that made @deepkit/types), Their rpc is way more performant.",
   },
   hapi: {
     hasRouter: true,
