+++ metrics/startup-fastify-routes.js
